[BE] 그룹장이 그룹을 나갈 경우에 대한 처리 추가

Closes #5

See merge request s08-final/S08P31C205!26

+++ backend/src/main/java/com/slembers/alarmony/alarm/service/GroupServiceImpl.java
@@ -53,10 +53,19 @@ public class GroupServiceImpl implements GroupService {
         Alarm alarm = alarmRepository.findById(groupId)
             .orElseThrow(() -> new CustomException(AlarmErrorCode.ALARM_NOT_FOUND));
 
+        if (member.equals(alarm.getHost()) && memberAlarmRepository.countByAlarmId(groupId) != 1) {
+            throw new CustomException(AlarmErrorCode.MEMBER_IN_GROUP);
+        }
+
         MemberAlarm memberAlarm = memberAlarmRepository.findByMemberAndAlarm(member, alarm)
             .orElseThrow(() -> new CustomException(AlarmErrorCode.MEMBER_NOT_IN_GROUP));
 
         alarmRecordRepository.deleteByMemberAlarm(memberAlarm);
         memberAlarmRepository.delete(memberAlarm);
+
+        if (member.equals(alarm.getHost())) {
+            alarmRepository.delete(alarm);
+        }
     }
+
 }
